==> Project Files Final/2B_Project_Regression_BayesianLinearmodels.Rmd <==
---
title: "INFSCI 2595 Final Project"
author: "Radhika Purohit"
subtitle: "Part 2B, Regression - Bayesian Linear models"
---

```{r setup, include=TRUE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(tidyverse)
theme_set(theme_linedraw())
library(rstanarm)
options(mc.cores = parallel::detectCores())
library(corrplot)
```


```{r}
df <- readr::read_csv("paint_project_train_data.csv", col_names = TRUE)
dfii <- df %>% 
  mutate(y = boot::logit( (response - 0) / (100 - 0) ) ) %>% 
  select(R, G, B, 
         Lightness, Saturation, Hue,
         y)

dfii %>% glimpse()
```

----
In this section, will use Bayesian approach to evaluate 2 models with relatively good performance from section 2A.  
have opted to use `rstanarm` for Bayesian analysis.  
----

## Bayesian Model Fitting

**Select 3 best models from 2A for Bayesian model fitting.**  

- Model `Model_7` is best model according to AIC/BIC.  
- Model `Model_5`is second best model according to AIC/BIC.  

# Reload Models
```{r, reload_mod01}
re_load_mod05 <- readr::read_rds("Model_5.rds")
re_load_mod07 <- readr::read_rds("Model_7.rds")
```


**1. Based on model `Model_05`**  
```{r model}
conflicted::conflicts_prefer(rstanarm::R2)
Mod_5 <- stan_lm(
  re_load_mod05$call$formula,
  data=dfii,
  prior = R2(location = 0.5),
  seed = 1324
)
```
**2. Based on model `Model_07`**  
```{r}
conflicted::conflicts_prefer(rstanarm::R2)
Mod_7 <- stan_lm(
  re_load_mod07$call$formula,
  data=dfii,
  prior = R2(location = 0.5),
  seed = 1324
)
```

```{r}
Mod_7 %>% readr::write_rds("Mod_7.rds")
Mod_5 %>% readr::write_rds("Mod_5.rds")
```
### Model Comparison

**Compare models based on information criterion.**  

```{r}
# Compute loo for each model
Mod_5_loo <- rstan::loo(Mod_5, k_threshold = 0.7)
Mod_7_loo <- rstan::loo(Mod_7, k_threshold = 0.7)
```

```{r}
# Add loo objects to the models
Mod_5$loo <- Mod_5_loo
Mod_7$loo <- Mod_7_loo

# Create stanreg_list with loo objects
my_models_loo <- stanreg_list(
  Mod_5, Mod_7,
  model_names = c("Model_5", "Model_7")
)

# Compare using loo_compare
loo_compare(my_models_loo)
```

The best model from the results seem to be Model_7.

### Examine Best Model

**Analyze the posterior behavior of the best model.**  

#### R-squared

```{r}
bayes_R2(Mod_7) %>% quantile(c(0.05, 0.5, 0.95))
```



#### Coefficients

**Visualize posterior regression coefficient plot.**  

```{r uncertainty in Model}
plot(Mod_7, pars = names(Mod_7$coefficients)) +
  geom_vline(xintercept = 0, color = "black", linetype = "dashed", size = 0.5) +
  geom_point(
    data = as_tibble(coef(re_load_mod07), rownames="coef"), aes(x = value, y = coef),
    shape = 5,
    size = 3,
    color = "purple"
  )
```

- The MLE estimated coefficients are plotted using purple diamonds.
- We see the prior constrained the coefficients.


#### Model Uncertainty

**Tasks**  

- Study the posterior uncertainty in the noise (residual error), 𝜎.
- How does the `lm()` maximum likelihood estimate (MLE) on 𝜎 relate to the posterior uncertainty on 𝜎 ?
- Do you feel the posterior is precise or are we quite uncertain about 𝜎

**Visualize and compare posterior distribution of σ across 3 models.**  

```{r}

# Assuming combined_posterior doesn't have a 'model' column
combined_posterior <- bind_rows(
  mutate(as.data.frame(Mod_7), model = "Mod_7"),
  mutate(as.data.frame(Mod_5), model = "Mod_5")
)

# Visualize posterior sigma with ggplot
combined_posterior %>%
  ggplot(aes(x = sigma, fill = model)) +
  geom_histogram(bins = 50, position = "identity", alpha = 0.7, color = "white") +
  geom_vline(aes(xintercept = sigma), linetype = "dotted", size = 0.5) +
  labs(title = "Posterior Sigma Comparison", x = "Sigma") +
  theme_minimal()

```

- The amount of uncertainty is varying across the two models, but almost similar, spanning the range of about 0.05 for model_7.  
- Thus the posterior σ is fairly precise as it does not span large range.  


**Compute the numbers on the best model.**  

```{r}
sprintf("Posterior mode σ: %.3f", sigma(Mod_7))
sprintf("MLE estimated σ: %.3f", sigma(re_load_mod07))
```

- For the best model identified as `Model_7`:
  - The posterior mode of σ from the Bayesian model is around 0.050.
  - The corresponding MLE estimated of σ is 0.045.  

----
