==> Project Files Final/3B_Classification-Bayesian-Models.Rmd <==
---
title: "INFSCI 2595 Final Project"
subtitle: "Part 3B, Classification - Bayesian Analysis"
author: "Radhika Purohit"
output: html_document
---

```{r}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
theme_set(theme_linedraw())
library(rstanarm)
options(mc.cores = parallel::detectCores())
library(corrplot)
```

```{r, read_final_data}
df <- readr::read_csv("paint_project_train_data.csv", col_names = TRUE)
```

----
In this section, will use Bayesian approach to evaluate the 3 baseline logistic models with relatively good performance from part 3A.  
----

## Preprocessing

The Binary output variable, `outcome`, is a numeric variable.  

```{r, show_outcome_class}
df %>% pull(outcome) %>% class()
```

However, there are **only** two unique values for `outcome`.  

```{r, show_outcome_values}
df %>% count(outcome)
```

```{r, make_iiiA_data}
dfiiiA <- df %>% 
  select(-response)

dfiiiA %>% glimpse()
```

----

## Bayesian Model Fitting

**Select 2 models from 3A for Bayesian model fitting.** 

- Based on the best BIC, select `mod_3A6`.  
- Based on the best AIC, select `mod_3A7`.  

```{r}
mod_3A6 <- readr::read_rds("mod_3A6.rds")
mod_3A7 <- readr::read_rds("mod_3A7.rds")
```


```{r model1}
mod_3B1_3A6 <- stan_glm(
  mod_3A6$call$formula,
  data = dfiiiA,
  family = binomial,
  prior = normal(0,1),
  seed = 1324
)
```



```{r model2}
mod_3B2_3A7 <- stan_glm(
  mod_3A7$call$formula,
  data = dfiiiA,
  family = binomial,
  prior = normal(0,1),
  seed = 1324
)
```


Save the 2 models to file.  

```{r}
mod_3B1_3A6 %>% readr::write_rds("mod_3B1_3A6.rds")
mod_3B2_3A7 %>% readr::write_rds("mod_3B2_3A7.rds")
mod_3A7 %>% readr::write_rds("mod_3A7.rds")
mod_3A6 %>% readr::write_rds("mod_3A6.rds")
```

----


### Model Comparison

**Compare models based on information criterion.**  

```{r}
mod_3B1_3A6$waic <- waic(mod_3B1_3A6)
mod_3B2_3A7$waic <- waic(mod_3B2_3A7)
```

```{r}
my_models <- stanreg_list(
  mod_3B1_3A6, mod_3B2_3A7,
  model_names = c("mod_3B1_3A6","mod_3B2_3A7"))
loo_compare(my_models, criterion = "waic")
```
- The model, mod_3B1_3A6, should be picked as the best model.  

### Best Model

**Analyze the posterior behavior of the best model.**  

#### Coefficients

**Visualize posterior regression coefficient plot.**  

```{r Uncertainty}
plot(mod_3B1_3A6, pars = names(mod_3B1_3A6$coefficients)) +
  geom_vline(xintercept = 0, color = "black", linetype = "dashed", size = 0.5) +
  geom_point(
    data = as_tibble(coef(mod_3A6), rownames="coef"), aes(x = value, y = coef),
    shape = 5,
    size = 3,
    color = "purple"
    )
```

- The MLE estimated coefficients are plotted using purple diamonds. 
- We see the prior slightly constrained most of the coefficients.  

**Check the correlation of the posterior coefficients.**  

```{r Correlations}
mod_3B1_3A6 %>% as_tibble() %>%
  select(all_of(names(mod_3B1_3A6$coefficients))) %>%
  cor() %>% corrplot(method="color", type="upper")
```

----
